pyproject.toml
[tool.poetry]
name = "scraper"
version = "0.1.0"
description = "RAG-based Web Scraping and Q&A Application"
license = "MIT"
authors = ["arkeodev <arkeodev@gmail.com>"]
packages = [
    { include = "scraper" }
]

[tool.poetry.scripts]
scraper = 'scraper.main:main'

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
pandas = "^2.0.2"
pydantic = "^1.10.7"
streamlit = "^1.25.0"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.3"
selenium = "^4.11.1"
webdriver-manager = "^4.0.1"
faiss-cpu = "^1.7.2"
openai = "^0.27.2"
transformers = "^4.41.2"
python-dotenv = "^1.0.1"
torch = "2.2.2"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
black = "^23.3.0"
pre-commit = "^3.2.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.core.masonry.api"
